python-oauth2/pyproject.toml
[project]
name = "python-oauth2"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyjwt>=2.9.0",
    "requests>=2.32.3",
]
